refactor: service 클래스 service 패키지 내부로 이동
