+++ src/main/java/roomescape/application/service/ReservationTimeService.java
@@ -1,4 +1,4 @@
-package roomescape.application;
+package roomescape.application.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
