+++ src/main/java/roomescape/presentation/web/controller/ReservationController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.application.ReservationService;
+import roomescape.application.service.ReservationService;
 import roomescape.domain.Reservation;
 import roomescape.presentation.web.request.CreateReservationWebRequest;
 import roomescape.presentation.web.response.ReservationWebResponse;
