+++ src/main/java/roomescape/presentation/web/controller/ReservationTimeController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.application.ReservationTimeService;
+import roomescape.application.service.ReservationTimeService;
 import roomescape.domain.ReservationTime;
 import roomescape.presentation.web.request.CreateReservationTimeWebRequest;
 import roomescape.presentation.web.response.ReservationTimeWebResponse;
